update: basic item checklist tool
