+++ src/app/tools/layout.tsx
@@ -1,0 +1,3 @@
+export default function ToolLayout({ children }: { children: React.ReactNode }) {
+	return <article className="w-full p-4">{children}</article>;
+}
